Encode month availability in a single string
Instead of having month availability encoded in a different tag for each
month in the record, encode it in a single tag with a string of 12
chars, which for now are just 0 (not available) and 1 (available), but
in the future there could be more keys, for example to specify that it
is only available during the first half of that month.

This not only reduces the size of the XML but it also makes parsing the
availabilities easier.