Add unique identifiers to records
To be able to uniquely identify a collectible, a unique identifier is
added to the record by the <id> tag. With this, we can for example use
Android's shared preferences (which is just a dictionary) to store which collectibles we have captured/donated by using {id: state} pairs.

An offset has been given to each collectible type and an id is given by
counting from 1 when parsing the data. The assumption that there are no
collections with over 100 elements has been made.