Get full size image instead of downscaled one
When the parser goes through the table of resources, it fetches the
image of the resource from src, which unfortunately points to
dynamically downscaled version and thus is not the highest quality
available.

With this commit, the src url is modified to fetch the full HD version
of the image.